+++ db/db_impl.h
@@ -341,9 +341,7 @@ class DBImpl : public DB {
 
   Cache* TEST_table_cache() { return table_cache_.get(); }
 
-  const WriteController& TEST_write_controler() const {
-    return write_controller_;
-  }
+  WriteController& TEST_write_controler() { return write_controller_; }
 
 #endif  // NDEBUG